Add enemy split

+++ src/ui/fps.py
@@ -8,11 +8,11 @@ settings = Settings()
 from src.ui.label import Label
 
 class FPS(Label):
-    def __init__(self, fps: int = 0) -> None:
+    def __init__(self, fps: float = 0) -> None:
         super().__init__(f"FPS {fps:5.2f}", settings.FPS_FONT, settings.FPS_SIZE,
                          True, settings.FPS_COLOR, settings.FPS_OPACITY,
                          settings.FPS_X, settings.FPS_Y + settings.FPS_Y_PADDING)
 
-    def update(self, fps: int) -> None:
+    def update(self, fps: float) -> None:
         self.text = f"FPS {fps:5.2f}"
         super().update()
